feat: separate perf tests
