+++ package.json
@@ -15,6 +15,7 @@
         "test:clear": "jest --clearCache",
         "test": "jest --detectOpenHandles",
         "test:coverage": "jest --coverage",
+        "test:perf": "jest --config=jest-perf.config.js",
         "test:generate": "ts-node test/scripts/generatePerfTestData.ts",
         "build": "tsc",
         "generate-parser": "cd src/grammar-parser && jison grammar-parser.jison",
